UPSTREAM: <carry>: pkg/migrator: Log some errors on V(2)
Some errors that can happen when writing resources back during migration
are not particularly significant, yet they can cause a lot or log spam.

This change marks some errors coming from the try function to be logged
on V(2) so that there are not that many log messages. These include:

* conflict
* not found
* UID precondition failed (effectively a conflict)